Optimize commands page performances
