+++ frontend/src/Commands/CommandsContainer.js
@@ -1,24 +1,7 @@
 import { Commands } from './Commands';
 import { connect } from 'react-redux';
 import Actions from '../actions';
-
-const mapStateToProps = (state, ownProps) => {
-    const fetching = state.commands.fetching;
-    const commands = state.commands.ids.map(id => state.commands.commands[id]);
-    if(!commands || commands.length === 0) {
-        return { fetching };
-    }
-    const categories = commands.reduce( (acc, cur) => {
-        let category = cur.category in acc ? acc[cur.category] : { commands: [] };
-        category.commands.push(cur);
-        return {...acc, [cur.category]: category};
-    }, {});
-
-    return {
-        categories,
-        fetching,
-    };
-}
+import { commandsSelector } from './selectors';
 
 const addErrorNotification = (title, message) => Actions.Notifications.add(title, message, 'error');
 
@@ -27,4 +10,4 @@ const mapDispatchToProps = {
     refresh: Actions.Commands.get,
 };
 
-export const CommandsContainer = connect(mapStateToProps, mapDispatchToProps)(Commands);
+export const CommandsContainer = connect(commandsSelector, mapDispatchToProps)(Commands);
